feat: add navigation links and site pages
A new file NavbarLinks was added to implement the navigation menu for the `sugaming-site`.
This menu contains links to different parts of the site such as LoL, CS2, etc. sections.
Added placeholder pages for the links in the navigation bar.
Updated the Badge component to add a new badge variant for static secondary items.
The i18n JSON files were updated to include translatable texts for these additions.