feat: add team join request endpoint
Implemented a new API endpoint in CS2TeamsController to handle team join requests.
Added relevant validation checks to ensure user and team exist, the user is not already in the team,
and the user hasn't made a previous request.
Error messages were updated in the i18n files to support these checks.
Added two new methods to the users service - `getById` and `getByIdOrThrow`.