feat: handle full team case in cs2 teams
Added a new exception for joining a team that is already full team.
The error messages are localizable as they've been added to the 'errors.json' files.
Changes are reflected in the service files, on invite/request acceptance.
There the team member count is validated before a new member is added.