feat: add discord account to database
A new table "discord_accounts" is created to support the Discord account linking feature.
Added a definition of the DiscordAccount model in the Prisma schema.
Added of a foreign key constraint for "userId", binding it to the "users" table.
This structure enables user information from Discord to be stored and accessed the feature.