feat: discord account linking and user existence checks
Changed the `GET /api/v1/auth/login/discord` to handle both authenticated and unauthenticated users.
If an authenticated user opens the resource, the discord account is linked to their existing user.
If the account is already linked/the user has a discord account linked, an exception is thrown.
If an unauthenticated user logs in with discord, tokens are generated for the existing linked user.
If no user can be found with this discord account, throw an exception.
Added getByDiscordId, getByDiscordIdOrThrow and linkDiscordAccount methods the UsersService.
Added users-discord-account-already-linked.exception.ts and corresponding translations.
Changed the discord callback URL variable name.