feat: steam account linking and user existence checks
Changed the `GET /api/v1/auth/login/steam` to handle both authenticated and unauthenticated users.
If an authenticated user opens the resource, the steam account is linked to their existing user.
If the account is already linked/the user has a steam account linked, an exception is thrown.
If an unauthenticated user logs in with steam, tokens are generated for the existing linked user.
If no user can be found with this steam account, throw an exception.
Added getBySteamId, getBySteamIdOrThrow and linkSteamAccount methods the UsersService.
Added users-steam-account-already-linked.exception.ts and corresponding translations.
Changed the discord callback URL variable name.