feat: add discord login functionality
Implemented a new endpoint for Discord login in `sugaming-site` client.
The Discord login button redirects to the `sugaming-api` `GET /api/v1/auth/login/discord` endpoint.
The callback is set to the new endpoint in the `sugaming-site`.
The new endpoint gets tokens from the `sugaming-api` and sets them as cookies.
Added a new DTO for Discord login in the API in order for the `code` parameter to be documented.