feat(#85): Log In page (#91)
* feat: add form and input components from shadcn/ui
Add Form and Input components.

* feat: login form component
A new LoginForm component has been created with `email` and `password` fields.
This component uses the react-hook-form library and zod for schema validation.
Alongside the component, its respective Storybook file has also been added.

* feat: add toast notification component
Added toast notification components in the common UI library using the shadcn/ui library.
Introduced the toast reducer and the useToast hook to control toast behavior.
Exported as a client component.

* feat: credentials login action
Attached the server action to the login form.
If an error occurs during authentication, information toast is displayed.

* feat: resolve site login/refresh caching issues
Resolved issues with caching on login and token refreshing server actions.
Thank you, @KokosTech! 🥹

* feat: implement internationalization in login form
Added translations to the login form, enabling multilingual functionality.
The site login function now uses the appropriate API locale.
Translation strings are included both in Bulgarian and English.

* fix: login form proper zod error handling
Modified the login form, so that the validation errors are displayed.

* fix: console error due to unsupported navbar props

* feat: login page template
Updated the login page with a new design including a banner placeholder and terms agreement.
Adjusted the body height property.

* feat: add discord login functionality
Implemented a new endpoint for Discord login in `sugaming-site` client.
The Discord login button redirects to the `sugaming-api` `GET /api/v1/auth/login/discord` endpoint.
The callback is set to the new endpoint in the `sugaming-site`.
The new endpoint gets tokens from the `sugaming-api` and sets them as cookies.
Added a new DTO for Discord login in the API in order for the `code` parameter to be documented.

* feat: discord auth error handling
Added error handling in cases, where the Discord login route returns an error.
If an error occurs, the client is redirected to the login page, and
the error is displayed as a toast notification.

* feat: update login page with translations and new icons
Added internationalization to the login page.
Added new login instructions and welcome messages.
Set a placeholder to the password field.
Added a new dependency, react-icons, for loading various icons.

* feat: add steam login support
Introduced support for logging in with Steam from the `sugaming-site`.
Created a new API DTO for the Steam login query and adjusted the existing DTO for Discord.
Modified the API Auth controller to include information about the Steam query parameters.

* feat: enable account linking in `sugaming-site`
Included the current client authentication in both the Steam and Discord API calls.
This ensures that if the API endpoints are called with a logged-in account, the account is linked.

* feat: update login form with a signup link
A signup link for users without an account has been added to the form for better accessibility.
The href values for Discord and Steam login methods have been updated to use environment variables.

* fix: storybook dependency errors + login form args
Resolved various issues preventing storybooks from starting.
Resolved mismatched arguments for the Site/LoginForm component.