feat: add api file upload for university proof images
This update introduces a file upload feature for the university proof images during user sign-up.
With the help of multer module, images uploaded are stored in a specified destination and filename.
A new 'multer' configuration has been added to lib config to specify the destination of the uploads.
Dependencies on the multer module and related types have also been added to the project.