feat: integrate ui with onboarding completion
The function completeOnboarding() was added to the API client which signals the completion of the
user's onboarding process.
In addition, a function to execute completeOnboarding was integrated into the onboarding-dialog UI.
If an error occurs during execution, a toast notification will appear with a corresponding error
message, otherwise, the dialog will close.