feat: add team verification
Reduced the maximum length for user's first and last names from 50 to 40.
Increased the minimum number of members in a CS2 team from 1 to 5 and introduced SU threshold.
Introduced a new utility function isCs2TeamVerified to filter teams based on verification status.
Updated the display of CS2 teams on the website to separate between verified and unverified teams.