feat: add logo component and adjust usage across various files
A new Logo component has been added in this commit, and it's being used in navbar, login, and layout
pages replacing previous text or iconic representations.
A few components have been reorganized for the Team Card component, and the export for the new Logo
component has been included in the server.ts file.