feat: add cs2 team page in sugaming-site
A new file has been created for individual team pages in the CS2 section of the `sugaming-site`.
This page imports and uses various components and services, and includes the team's name.
The code also incorporates a temporary solution for retrieving teams, which will need to be replaced
with a function to get specific teams by their IDs.