fix: update user service to include cs2 team details + wrong path parameter name
The change updates the user service to return all details of the user's Counter-Strike 2 team.
The DTO is also modified to provide a reference to the CS2 team details.
Additionally, the API end point is updated to take the inviteeId as a parameter when inviting a user
to a CS2 team.