refactor: team member removal validation
Refactored the validation logic when removing members from a team in the cs2-teams service.
It now validates firstly if the user exists and is part of the team, and then checks if the user is
either the captain or if they are trying to remove themselves.
This allows users themselves to leave the team.